// src/apis/user/user.enum.ts
export enum UserType {
  ROOT = 'Root',
  ADMIN = 'Admin',
  HR = 'HR',
  User = 'User',
  CANDIDATE = 'Candidate',
  EMPLOYEE = 'Employee', // ứng viên chấp nhận vào làm việc thì sẽ chuyển từ Candidate sang Employee
  C_B = 'C&B',
}